feat: user can add new category.
